+++ test/registered/unit/test_model_overrides.py
@@ -15,6 +15,7 @@ from types import SimpleNamespace
 from typing import Optional
 from unittest.mock import patch
 
+from sglang.srt.arg_groups import attention_hook
 from sglang.srt.arg_groups import overrides as overrides_module
 from sglang.srt.arg_groups.arg_utils import A, Arg, resolvable_fields
 from sglang.srt.arg_groups.overrides import (
@@ -357,14 +358,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             enable_dp_attention=enable_dp_attention,
             enable_hierarchical_cache=enable_hierarchical_cache,
         )
-        args.is_attention_backend_not_set = lambda: all(
-            backend is None
-            for backend in (
-                args.attention_backend,
-                args.prefill_attention_backend,
-                args.decode_attention_backend,
-            )
-        )
         mixer_types = []
         if sparse_attention:
             mixer_types.append("minicpm4")
@@ -456,7 +449,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             disaggregation_mode="null",
             enable_dp_attention=False,
             enable_hierarchical_cache=False,
-            is_attention_backend_not_set=lambda: True,
         )
         config = SimpleNamespace(
             has_minicpm_sparse_attention=True,
@@ -630,7 +622,11 @@ class TestGoldenModelOverrides(_IsolatedPublish):
 
     def test_minimax_m2_sm10x_nvfp4_uses_routed_trtllm(self):
         """MiniMax-M2 NVFP4 auto must avoid the unsupported plain TRT-LLM path."""
-        with patch.object(overrides_module, "is_sm100_supported", return_value=True):
+        # Every module that asks: the attention handler validates what the
+        # override family picks, and each holds its own import.
+        with patch.object(
+            overrides_module, "is_sm100_supported", return_value=True
+        ), patch.object(attention_hook, "is_sm100_supported", return_value=True):
             explicit = self._construct(
                 "MiniMaxM2ForCausalLM",
                 "llama",
@@ -765,7 +761,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
                 speculative_draft_attention_backend=None,
                 page_size=None,
                 mamba_radix_cache_strategy="auto",
-                is_attention_backend_not_set=lambda: True,
                 get_model_config=lambda: model_config,
             ),
             hf_config,
@@ -971,7 +966,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
         server_args.speculative_algorithm = "DFLASH"
         server_args.prefill_attention_backend = "triton"
         server_args.speculative_draft_attention_backend = "fa3"
-        server_args.is_attention_backend_not_set = lambda: False
 
         with (
             patch.object(overrides_module, "is_blackwell_supported", return_value=True),
@@ -1093,7 +1087,9 @@ class TestGoldenModelOverrides(_IsolatedPublish):
                 _gpt_oss_overrides(
                     SimpleNamespace(
                         dtype="float16",
-                        is_attention_backend_not_set=lambda: False,
+                        attention_backend="triton",
+                        prefill_attention_backend=None,
+                        decode_attention_backend=None,
                     ),
                     SimpleNamespace(architectures=["GptOssForCausalLM"]),
                 )
@@ -1608,11 +1604,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             )
             defaults.update(kw)
             args = SimpleNamespace(**defaults)
-            args.is_attention_backend_not_set = lambda: (
-                args.attention_backend is None
-                and args.prefill_attention_backend is None
-                and args.decode_attention_backend is None
-            )
             return args
 
         hf = _hf()
@@ -1952,11 +1943,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             )
             defaults.update(kw)
             ns = SimpleNamespace(**defaults)
-            ns.is_attention_backend_not_set = lambda: (
-                ns.attention_backend is None
-                and ns.prefill_attention_backend is None
-                and ns.decode_attention_backend is None
-            )
             ns.get_attention_backends = lambda: (
                 ns.prefill_attention_backend or ns.attention_backend,
                 ns.decode_attention_backend or ns.attention_backend,
@@ -2255,7 +2241,8 @@ class TestGoldenModelOverrides(_IsolatedPublish):
         def _args(default_backend, **kw):
             defaults = dict(
                 attention_backend=None,
-                _get_default_attn_backend=lambda **_: default_backend,
+                prefill_attention_backend=None,
+                decode_attention_backend=None,
                 use_mla_backend=lambda: False,
                 get_model_config=lambda: None,
                 mamba_radix_cache_strategy="auto",
@@ -2263,9 +2250,17 @@ class TestGoldenModelOverrides(_IsolatedPublish):
                 speculative_algorithm=None,
             )
             defaults.update(kw)
-            return SimpleNamespace(**defaults)
+            args = SimpleNamespace(**defaults)
+            args.default_backend_for_test = default_backend
+            return args
 
-        with patch.object(overrides_module, "is_sm100_supported", return_value=True):
+        with patch.object(
+            overrides_module, "is_sm100_supported", return_value=True
+        ), patch.object(
+            overrides_module,
+            "get_default_attn_backend",
+            lambda server_args, **_: server_args.default_backend_for_test,
+        ):
             # radix on + no extra buffer + no spec -> page_size=1 path
             self.assertEqual(
                 _qwen3_5_hybrid_overrides(_args("trtllm_mha"), None),
@@ -2422,11 +2417,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             )
             defaults.update(kw)
             ns = SimpleNamespace(**defaults)
-            ns.is_attention_backend_not_set = lambda: (
-                ns.attention_backend is None
-                and ns.prefill_attention_backend is None
-                and ns.decode_attention_backend is None
-            )
             return ns
 
         hf = SimpleNamespace()
@@ -2585,7 +2575,8 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             defaults = dict(
                 device="cuda",
                 attention_backend=None,
-                is_attention_backend_not_set=lambda: True,
+                prefill_attention_backend=None,
+                decode_attention_backend=None,
                 # keep the (now-absorbed) quant/moe blocks inert so these
                 # assertions stay attention-only
                 moe_runner_backend="triton",
@@ -2742,7 +2733,6 @@ class TestGoldenModelOverrides(_IsolatedPublish):
 
         def _args(**kw):
             defaults = dict(
-                is_attention_backend_not_set=lambda: True,
                 attention_backend=None,
                 prefill_attention_backend=None,
                 decode_attention_backend=None,
@@ -2869,7 +2859,9 @@ class TestGoldenModelOverrides(_IsolatedPublish):
             defaults = dict(
                 speculative_algorithm=None,
                 enable_hierarchical_cache=False,
-                is_attention_backend_not_set=lambda: False,
+                attention_backend="triton",
+                prefill_attention_backend=None,
+                decode_attention_backend=None,
             )
             defaults.update(kw)
             return SimpleNamespace(**defaults)
